Create data_visualization
creating a dashboard that displays the percentage of flights running under specific distance group. Distance group is the distance intervals, every 250 miles, for flight segment. If the flight covers to 500 miles, it will be under distance group 2 (250 miles + 250 miles).
this will we done using Pandas and Dash.
this project had been able to be made with tthe help of coursea